moved azure firewall stuff to common ext
