+++ lib/ansible/modules/cloud/azure/azure_rm_azurefirewall.py
@@ -362,7 +362,7 @@ etag:
 import time
 import json
 import re
-from ansible.module_utils.azure_rm_common import AzureRMModuleBase
+from ansible.module_utils.azure_rm_common_ext import AzureRMModuleBaseExt
 from copy import deepcopy
 try:
     from msrestazure.azure_exceptions import CloudError
@@ -378,7 +378,7 @@ class Actions:
     NoAction, Create, Update, Delete = range(4)
 
 
-class AzureRMAzureFirewalls(AzureRMModuleBase):
+class AzureRMAzureFirewalls(AzureRMModuleBaseExt):
     def __init__(self):
         self.module_arg_spec = dict(
             resource_group=dict(
@@ -892,149 +892,6 @@ class AzureRMAzureFirewalls(AzureRMModuleBase):
         except CloudError as e:
            return False
         return response.as_dict()
-
-    def inflate_parameters(self, spec, body, level):
-        if isinstance(body, list):
-            for item in body:
-                self.inflate_parameters(spec, item, level)
-            return
-        for name in spec.keys():
-            # first check if option was passed
-            param = body.get(name)
-            if not param:
-                continue
-            # check if pattern needs to be used
-            pattern = spec[name].get('pattern', None)
-            if pattern:
-                param = self.normalize_resource_id(param, pattern)
-                body[name] = param
-            disposition = spec[name].get('disposition', '*')
-            if level == 0 and not disposition.startswith('/'):
-                continue
-            if disposition == '/':
-                disposition = '/*'
-            parts = disposition.split('/')
-            if parts[0] == '':
-                # should fail if level is > 0?
-                parts.pop(0)
-            target_dict = body
-            while len(parts) > 1:
-                target_dict = target_dict.setdefault(parts.pop(0), {})
-            targetName = parts[0] if parts[0] != '*' else name
-            target_dict[targetName] = body.pop(name)
-            if spec[name].get('options'):
-                self.inflate_parameters(spec[name].get('options'), target_dict[targetName], level + 1)
-
-    def normalize_resource_id(self, value, pattern):
-        '''
-        Return a proper resource id string..
-
-        :param resource_id: It could be a resource name, resource id or dict containing parts from the pattern.
-        :param pattern: pattern of resource is, just like in Azure Swagger
-        '''
-        pattern_parts = pattern.split('/')
-        for i in range(len(pattern_parts)):
-            x = re.sub('[{} ]+', '', pattern_parts[i], 2)
-            if len(x) < len(pattern_parts[i]):
-                pattern_parts[i] = '{' + re.sub('([a-z0-9])([A-Z])', r'\1_\2', x).lower() + '}'
-
-        if isinstance(value, str):
-            value_parts = value.split('/')
-            if len(value_parts) == 1:
-                value_dict = {}
-                value_dict['name'] = value
-            else:
-                if len(value_parts) != len(pattern_parts):
-                    return None
-            value_dict = {}
-            for i in range(len(value_parts)):
-                if pattern_parts[i].startswith('{'):
-                    value_dict[pattern_parts[i][1:-1]] = value_parts[i]
-                elif value_parts[i].lower() != pattern_parts[i].lower():
-                    return None
-        elif isinstance(value, dict):
-            value_dict = value
-        else:
-            return None
-
-        if not value_dict.get('subscription_id'):
-            value_dict['subscription_id'] = self.subscription_id
-        if not value_dict.get('resource_group'):
-            value_dict['resource_group'] = self.resource_group
-
-        for i in range(len(pattern_parts)):
-            if pattern_parts[i].startswith('{'):
-                value = value_dict.get(pattern_parts[i][1:-1], None)
-                if not value:
-                    return None
-                pattern_parts[i] = value
-
-        return '/'.join(pattern_parts)
-
-    def create_compare_modifiers(self, arg_spec, path, result):
-        for k in arg_spec.keys():
-            o = arg_spec[k]
-            updatable = o.get('updatable', True)
-            comparison = o.get('comparison', 'default')
-            p = (path +
-                 ('/' if len(path) > 0 else '') +
-                 o.get('disposition', '*').replace('*', k) +
-                 ('/*' if o['type'] == 'list' else ''))
-            if comparison != 'default' or not updatable:
-                result[p] = { 'updatable': updatable, 'comparison': comparison }
-            if o.get('options'):
-                self.create_compare_modifiers(o.get('options'), p, result)
-
-    def default_compare(self, modifiers, new, old, path, result):
-        if new is None:
-            return True
-        elif isinstance(new, dict):
-            if not isinstance(old, dict):
-                result['compare'] = 'changed [' + path + '] old dict is null'
-                return False
-            for k in new.keys():
-                if not self.default_compare(modifiers, new.get(k), old.get(k, None), path + '/' + k, result):
-                    return False
-            return True
-        elif isinstance(new, list):
-            if not isinstance(old, list) or len(new) != len(old):
-                result['compare'] = 'changed [' + path + '] length is different or null'
-                return False
-            if isinstance(old[0], dict):
-                key = None
-                if 'id' in old[0] and 'id' in new[0]:
-                    key = 'id'
-                elif 'name' in old[0] and 'name' in new[0]:
-                    key = 'name'
-                else:
-                    key = old[0].keys()[0]
-                new = sorted(new, key=lambda x: x.get(key, None))
-                old = sorted(old, key=lambda x: x.get(key, None))
-            else:
-                new = sorted(new)
-                old = sorted(old)
-            for i in range(len(new)):
-                if not self.default_compare(modifiers, new[i], old[i], path + '/*', result):
-                    return False
-            return True
-        else:
-            updatable = modifiers.get(path, {}).get('updatable', True)
-            comparison = modifiers.get(path, {}).get('comparison', 'default')
-            if path == '/location' or path.endswith('locationName'):
-                new = new.replace(' ', '').lower()
-                old = old.replace(' ', '').lower()
-            elif path.endswith('adminPassword') or path.endswith('administratorLoginPassword') or path.endswith('createMode'):
-                return True
-            if str(new) == str(old):
-                result['compare'] = result.get('compare', '') + "(" + str(new) + ":" + str(old) + ")"
-                return True
-            else:
-                result['compare'] = 'changed [' + path + '] ' + str(new) + ' != ' + str(old)
-                if updatable:
-                    return False
-                else:
-                    # XXX change new value to old
-                    self.module.warn("property '" + path + "' cannot be updated (" + str(old) + "->" + str(new) + ")")
 
 
 def main():
